culminando excepciones

+++ src/main/java/com/example/clinica/service/impl/PacienteService.java
@@ -2,8 +2,12 @@ package com.example.clinica.service.impl;
 
 
 import com.example.clinica.entity.Paciente;
+import com.example.clinica.exception.BadRequestException;
+import com.example.clinica.exception.ResourceNotFoundException;
 import com.example.clinica.repository.IPacienteRepository;
 import com.example.clinica.service.IPacienteService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,7 +16,9 @@ import java.util.Optional;
 
 @Service
 public class PacienteService implements IPacienteService {
+    //Excepciones en guardar, buscar por id y eliminar
 
+    public static final Logger logger = LoggerFactory.getLogger(PacienteService.class);
     private IPacienteRepository pacienteRepository;
 
     public PacienteService(IPacienteRepository pacienteRepository) {
@@ -21,7 +27,13 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public Paciente guardarPaciente(Paciente paciente) {
-        return pacienteRepository.save(paciente);
+        try {
+            logger.info("Inicia el guardado del paciente");
+            return pacienteRepository.save(paciente);
+        }catch (RuntimeException e){
+            logger.error("Error guardando el paciente " + e);
+            throw new BadRequestException("El paciente no pudo ser guardado");
+        }
     }
 
     @Override
@@ -41,7 +53,13 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public void eliminarPaciente(Integer id) {
-        pacienteRepository.deleteById(id);
+        Optional<Paciente> pacienteEncontrado = buscarPorId(id);
+        if (pacienteEncontrado.isPresent()){
+            pacienteRepository.deleteById(id);
+        }else {
+            throw new ResourceNotFoundException("El paciente no ha sido encontrado");
+        }
+
     }
 
     @Override
@@ -53,6 +71,5 @@ public class PacienteService implements IPacienteService {
     public List<Paciente> buscarPorUnaParteApellido(String parte){
         return pacienteRepository.buscarPorParteApellido(parte);
     }
-
 
 }
